Refactor: Move ReactDOMServer into own file
